% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loudly.R
\name{bind_loudly}
\alias{bind_loudly}
\title{Evaluate a decorated function}
\usage{
bind_loudly(.l, .f, ...)
}
\arguments{
\item{.l}{A loud value (a list of two elements)}

\item{.f}{A loud function to apply to the returning value of .l}

\item{...}{Further parameters to pass to .f}
}
\value{
A list with elements .f(.l$result) and concatenated logs.
}
\description{
Evaluate a decorated function
}
\examples{
loud_sqrt <- loudly(sqrt)
loud_exp <- loudly(exp)
3 |> loud_sqrt() |> bind_loudly(loud_exp)
}
