% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loudly.R
\name{flat_loudly}
\alias{flat_loudly}
\title{Evaluate a non-loud function on a loud value}
\usage{
flat_loudly(.l, .f, ...)
}
\arguments{
\item{.l}{A loud value (a list of two elements)}

\item{.f}{A non-loud function}

\item{...}{Further parameters to pass to .f}
}
\value{
Returns the result of .f(.l$value)
}
\description{
Evaluate a non-loud function on a loud value
}
\examples{
loud_value(3) |> flat_loudly(sqrt)
}
