% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loudly.R
\name{loud_value}
\alias{loud_value}
\title{Create a loud value}
\usage{
loud_value(.x)
}
\arguments{
\item{.x}{Any object}
}
\value{
Returns a loud value with the object as the $result
}
\description{
Create a loud value
}
\examples{
loud_value(3)
}
