% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loudly.R
\name{pick}
\alias{pick}
\title{Retrieve an element from a loud value}
\usage{
pick(.l, .e)
}
\arguments{
\item{.l}{A loud value}

\item{.e}{Element of interest to retrieve, one of "value" or "log"}
}
\value{
The \code{value} or \code{log} element of the loud value .l
}
\description{
Retrieve an element from a loud value
}
\examples{
loud_sqrt <- loudly(sqrt)
loud_exp <- loudly(exp)
3 |> loud_sqrt() \%>=\% loud_exp() |> pick("value")
}
