% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loudly.R
\name{purely}
\alias{purely}
\title{Capture all errors, warnings and messages}
\usage{
purely(.f)
}
\arguments{
\item{.f}{A function to decorate}
}
\value{
A function which returns a list. The first element of the list, $value, is the result of
the original function .f applied to its inputs. The second element, $log is NULL in case everything
goes well. In case of error/warning/message, $value is NA and $log holds the message.
purely() is used by loudly() to allow the latter to handle errors.
}
\description{
Capture all errors, warnings and messages
}
\examples{
purely(log)(10)
purely(log)(-10)
}
