% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loudly.R
\name{loudly}
\alias{loudly}
\title{Add a simple logging message to any function}
\usage{
loudly(.f)
}
\arguments{
\item{.f}{A function to decorate}
}
\value{
A function which returns a list. The first element of the list, $result, is the result of
the original function .f applied to its inputs, and the second element is a log message, $log.
}
\description{
Add a simple logging message to any function
}
\examples{
loudly(sqrt)(10)
}
